Cleanup database subclassing
